added new function - clean_tweet_text_meta
added new in-line function to remove meta data from tweet text before writing to csv data file. this function also captures the meta data and outputs it to a csv file accompaning the main tweet text csv. the id field links the two data sources together. the intent is to be able to web scrab the URL data associated to the tweet text for a richer insight into the meaning of the tweet.